requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
multidict==4.7.6
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.4.2

setuptools~=47.1.1
boto3~=1.13.26
requests~=2.23.0
botocore~=1.16.26